\name{CAZy}
\alias{CAZy}
\title{
Predict carbohydrate metabolism
}
\description{
The function mappes the species in your samples to the CAZy database and predicts the abundance of carbohydrate-processessing genes. You should run this in the same directory where the CAZy data base files are.
}
\usage{
CAZy(species)
}

\arguments{
  \item{species}{
The species-level taxonomic table.
}
}


\references{
Cantarel BL, Coutinho PM, Rancurel C, Bernard T, Lombard V, Henrissat B (2009) The Carbohydrate-Active EnZymes database (CAZy): an expert resource for Glycogenomics. Nucleic Acids Res 37:D233-238 [PMID: 18838391].
http://www.cazy.org/
}
\author{
Katri Korpela
}


