\name{mareGUI}
\alias{mreGUI}
\title{
Graphical User Interface for mare
}
\description{
Allows the user to interact with mare using a GUI instead of writing R script. Do not use both GUI and R command line at the same time! Close the GUI if you want to run script directly in R.
}
\usage{
mareGUI()
}

\author{
Katri Korpela
}

\examples{
\dontrun{
mareGUI()
}
}

