\name{CAZy}
\alias{CAZy}
\title{
Predict carbohydrate metabolism
}
\description{
The function mappes the taxa in your samples to the CAZy database and predicts the abundance of taxa producing carbohydrate-active enzymes. 
}
\usage{
CAZy(genus.table)
}

\arguments{
  \item{genus.table}{
The organised genus-level taxonomic table.
}
}


\references{
Cantarel BL, Coutinho PM, Rancurel C, Bernard T, Lombard V, Henrissat B (2009) The Carbohydrate-Active EnZymes database (CAZy): an expert resource for Glycogenomics. Nucleic Acids Res 37:D233-238 [PMID: 18838391].
http://www.cazy.org/
}
\author{
Katri Korpela
}


